fact-generator/requirements.txt
Faker==28.4.1
Flask==3.0.3
opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
openai==1.70.0
requests==2.32.3
# httpx==0.27.2 # downgrade for {"error":"Client.__init__() got an unexpected keyword argument 'proxies'"} error